Move is_ebs_volume_encrypted to helpers.py
