+++ aws/ec2/test_ec2_ebs_volume_encrypted.py
@@ -1,29 +1,8 @@
 import pytest
 
 from aws.ec2.resources import ec2_ebs_volumes
+from aws.ec2.helpers import is_ebs_volume_encrypted
 
-def is_ebs_volume_encrypted(ebs):
-    """
-    Checks the EBS volume 'Encrypted' value.
-
-    >>> is_ebs_volume_encrypted({'Encrypted': True})
-    True
-    >>> is_ebs_volume_encrypted({'Encrypted': False})
-    False
-    >>> is_ebs_volume_encrypted({})
-    Traceback (most recent call last):
-    ...
-    KeyError: 'Encrypted'
-    >>> is_ebs_volume_encrypted(0)
-    Traceback (most recent call last):
-    ...
-    TypeError: 'int' object is not subscriptable
-    >>> is_ebs_volume_encrypted(None)
-    Traceback (most recent call last):
-    ...
-    TypeError: 'NoneType' object is not subscriptable
-    """
-    return ebs['Encrypted']
 
 @pytest.mark.ec2
 @pytest.mark.parametrize(
